test: continue splitting modbus transport tests by area
